requirements.txt
starlette>=0.30.0
pydantic>=2
rich>=13
pytest>=8
torch>=2
numpy>=1
setuptools>=68
bittensor>=8.5.1
python-socks>=2.4.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
pyjwt>=2.9.0
fastapi>=0.68.0
uvicorn>=0.15.0